requirements.txt
boto3==1.24.11
pandas==1.4.2
openpyxl==3.0.10
awscli==1.18.69
python-dotenv==0.20.0